changed file/folder structure for components with stylesheets for consistency
