+++ client/components/ClockOutPortal/ClockOutPortal.css
@@ -1,0 +1,96 @@
+.card {
+  display: flex;
+  flex-direction: column;
+  height: 90%;
+  justify-content: center;
+  align-items: center;
+  height: 400px;
+  /* filter: drop-shadow(0 0 12px #08622778); */
+  box-shadow: 0 0 6px rgb(47, 112, 47);
+  border-radius: 10px;
+  margin: 0 auto;
+  border: 2px solid rgba(7, 67, 7, 0.497);
+  width: 65vh;
+  padding-top: 50px;
+}
+
+.card-body {
+  flex-grow: 1;
+}
+
+.card-title {
+  font-size: 1.5rem;
+  font-weight: bold;
+  margin-bottom: 1rem;
+}
+
+.card-text {
+  margin-bottom: 1rem;
+  padding-bottom: 50px;
+  font-size: 20px;;
+}
+
+.card-buttons {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: auto;
+  padding-top: 50px;
+}
+
+.card-buttons button {
+  margin-left: 0.5rem;
+  margin-right: 0.5rem;
+}
+
+.btn-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 18%;
+  padding: 20px 0
+} 
+
+/* Font Awesome Icon color */
+.button .icon-white path {  
+  fill: white; 
+}
+
+.clock {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  width: 42%;
+  padding-bottom: 20px;
+  margin: auto;
+  object-fit: contain;
+}
+
+.clock__date,
+.clock__time {
+  font-size: 30px;
+  display: inline-block;
+  margin: 0;
+  text-shadow: 0 1px 0 #ccc, 0 2px 0 #c9c9c9, 0 0 5px rgba(29, 163, 60, 0.1),
+    0 1px 3px rgba(0, 0, 0, 0.3), 0 2px 2px rgba(0, 0, 0, 0.15);
+  background-color: #075322dc;
+  -webkit-background-clip: text;
+  -moz-background-clip: text;
+  background-clip: text;
+  color: transparent;
+  text-shadow: rgba(255, 255, 255, 0.5) 0px 3px 3px;
+}
+
+/* .clock_in_portal {
+  margin-top: 1%;
+  display: flex;
+  grid-template-rows: 1fr;
+  justify-content: space-around;
+}
+
+.clock_out_portal {
+  margin-top: 1%;
+  display: flex;
+  grid-template-rows: 1fr;
+  justify-content: space-around;
+} */
