+++ client/components/pinpad/BackButton/BackButton.css
@@ -1,0 +1,15 @@
+.back_button {
+  width: 100%;
+  height: 100px;
+  background: #f4cf003e;
+  box-shadow:  0 0 6px rgba(226, 201, 10, 0.373);
+  background: rgb(255, 122, 117);
+  backdrop-filter: blur(4px);
+  border-radius: 14px;
+  font-size: 2.5rem;
+  border: 1px solid rgb(255, 122, 117);
+}
+
+.back_button:hover {
+  background-color: rgba(246, 163, 160, 0.715);
+}
