+++ client/components/pinpad/NumButton/NumButton.css
@@ -1,0 +1,17 @@
+.num_button {
+  width: 100%;
+  height: 100px;
+  /* background-color: #f4cf009f; */
+  border-radius: 10px;
+  font-size: 3rem;
+  background: #f4cf003e;
+  box-shadow:  0 0 6px rgba(226, 201, 10, 0.373);
+  backdrop-filter: blur(4px);
+  -webkit-backdrop-filter: blur(4px);
+  border-radius: 14px;
+  border: 1px solid rgba(209, 183, 12, 0.597);
+}
+
+.num_button:hover {
+  background-color: rgba(249, 246, 229, 0.881);
+}
