+++ client/components/pinpad/NumButton/NumButton.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './NumButton.css';
 
 type Props = {
   clicked: () => void
