+++ client/components/pinpad/PinDisplay/PinDisplay.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './PinDisplay.css';
 
 type Props = {
   val: string
@@ -6,7 +7,8 @@ type Props = {
 const PinDisplay: React.FC<Props> = ({ val }) => {
 
   return (
-    <div className='pin-display pin-pad'>
+    // changed this className to pin-display to reflect the css file
+    <div className='pin-display'>
       <h1>{val}</h1>
     </div>
   )
