+++ client/components/pinpad/PinPad/PinPad.css
@@ -1,0 +1,22 @@
+.pinpad_container {
+  border-radius: 10px;
+  margin: 0 auto;
+  border: 2px solid rgba(6, 121, 6, 0.497);
+  box-shadow: 0 0 6px rgb(47, 112, 47);
+  width: 44%;
+  margin-bottom: 30px;
+  padding-bottom: 30px;
+  padding-top: 30px;
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  filter: drop-shadow(0 0 15px #9ea7a951);
+}
+
+.pinpad {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  padding: 10px;
+  width: 70%;
+}
