--- client/components/pinpad/styles.pinpad.css
@@ -1,104 +1,0 @@
-.pinpad_container {
-  border-radius: 10px;
-  margin: 0 auto;
-  border: 2px solid rgba(6, 121, 6, 0.497);
-  box-shadow: 0 0 6px rgb(47, 112, 47);
-  width: 44%;
-  margin-bottom: 30px;
-  padding-bottom: 30px;
-  padding-top: 30px;
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  filter: drop-shadow(0 0 15px #9ea7a951);
-}
-
-.pinpad {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  padding: 10px;
-  width: 70%;
-}
-
-.pin-display {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 1px solid rgba(7, 67, 7, 0.594);
-  box-shadow: 0 0 4px rgb(4, 4, 4);
-  box-shadow: 0 0 4px rgb(47, 112, 47);
-  margin-top: 10px;
-  margin-bottom: 25px;
-  padding: 5px;
-  width: 140px;
-  border-radius: 10px;
-  height: 59px;
-  width: 160px;
-}
-
-.num_button {
-  width: 100%;
-  height: 100px;
-  /* background-color: #f4cf009f; */
-  border-radius: 10px;
-  font-size: 3rem;
-  background: #f4cf003e;
-  box-shadow:  0 0 6px rgba(226, 201, 10, 0.373);
-  backdrop-filter: blur(4px);
-  -webkit-backdrop-filter: blur(4px);
-  border-radius: 14px;
-  border: 1px solid rgba(209, 183, 12, 0.597);
-}
-
-.num_button:hover {
-  background-color: rgba(249, 246, 229, 0.881);
-}
-
-.back_button:hover {
-  background-color: rgba(246, 163, 160, 0.715);
-}
-
-.submit_btn:hover {
-  background-color: rgba(210, 250, 205, 0.698);
-}
-
-/* .back_button {
-  width: 100%;
-  height: 100px;
-  background-color: rgba(255, 122, 117, 0.793);
-  border-radius: 10px;
-  font-size: 2.5rem;
-  background: rgb(255, 122, 117);
-  box-shadow: 0 8px  0 rgb(251, 145, 141);
-  backdrop-filter: blur(4px);
-  -webkit-backdrop-filter: blur(4px);
-  border-radius: 14px;
-  border: 1px solid rgb(142, 189, 231);
-} */
-
-.submit_btn {
-  width: 100%;
-  height: 100px;
-  border-radius: 15px;
-  font-size: 2.5rem;
-  background: rgba(12, 143, 15, 0.947);
-  box-shadow: 0 1px 2px 0 rgba(42, 142, 44, 0.825);
-  backdrop-filter: blur(4px);
-  -webkit-backdrop-filter: blur(4px);
-  border-radius: 14px;
-  border: 1px solid rgb(142, 231, 163);
-}
-
-.back_button {
-  width: 100%;
-  height: 100px;
-  background: #f4cf003e;
-  box-shadow:  0 0 6px rgba(226, 201, 10, 0.373);
-  background: rgb(255, 122, 117);
-  backdrop-filter: blur(4px);
-  border-radius: 14px;
-  font-size: 2.5rem;
-  border: 1px solid rgb(255, 122, 117);
-
-}
